v.1.0.3 PROTOTYPE VERSION, Minor Revision 3
-= FIXED =-

* Fixed the error that the SMS adapter throws exception when rendering
the listview upon scrolling up or down when there are many SMS in a
conversation thread. This might be due to the view reuse.  As a
workaround solution, view reuse is always ignored and will always
inflate the view. This
can be recreated by going inside a conversation containing many sms and
then scroll it upwards. (SMSAdapter.java line 60 - 80)